requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.3
django-dotenv==1.4.2
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
Faker==15.3.4
filelock==3.8.1
flake8==6.0.0
identify==2.5.9
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pre-commit==2.20.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
toposort==1.7
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.0
